+++ packages/react-core/src/operations/constants/SpatialIndexTypes.js
@@ -2,5 +2,6 @@ export const SpatialIndex = Object.freeze({
   H3: 'h3',
   S2: 's2',
   QUADKEY: 'quadkey',
-  QUADINT: 'quadint'
+  QUADINT: 'quadint',
+  QUADBIN: 'quadbin'
 });